+++ app/views/layouts/weight_records.html.erb
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>StudyApp</title>
+  <%= csrf_meta_tags %>
+  <%= stylesheet_link_tag    'application', media: 'all' %>
+  <%= javascript_include_tag 'application' %>
+</head>
+
+<body>
+<div id="<%= "#{controller.controller_name}-#{controller.action_name}" %>">
+  <%= yield %>
+</div>
+</body>
+</html>
