+++ app/views/weight_records/index.html.erb
@@ -1,1 +1,6 @@
-<%= javascript_pack_tag "weight_record" %>
+<app-header
+  heading-text="体重計アプリ"
+></app-header>
+<input-record></input-record>
+<line-chart></line-chart>
+<%= javascript_pack_tag "weight_record/weight_record" %>
